Fixed some setup bugs in the Session class.

+++ ManuvrOS/Drivers/INA219/INA219.cpp
@@ -238,7 +238,7 @@ bool INA219::process_read_data(void) {
 		markRegRead(INA219_REG_SHUNT_VOLTAGE);
 
           StringBuilder output;
-          output.concatf("%.3f\t %.3f\t %.3f\t %.3f\n", local_shunt, local_bus, local_current, local_power);
+          output.concatf("%.3f\t %.3f\t %.3f\t %.3f\n", (double) local_shunt, (double) local_bus, (double) local_current, (double) local_power);
           StaticHub::log(&output);
 		return true;
 	}
